chore: add helper for assertion on liquidity strategy types

+++ test/v3/LiquidityStrategy/LiquidityStrategy.t.sol
@@ -137,7 +137,7 @@ contract LiquidityStrategy_Test is LiquidityStrategy_BaseTest {
 
     // Should be expansion: add debt (token0), take collateral (token1)
     assertEq(ctx.prices.poolPriceAbove, true);
-    assertEq(uint256(action.dir), uint256(LQ.Direction.Expand));
+    assertEq(action.dir, LQ.Direction.Expand);
     assertTrue(action.amountOwedToPool > 0); // Debt in
     assertEq(action.amount0Out, 0); // No debt out
     assertTrue(action.amount1Out > 0); // Collateral out
@@ -156,7 +156,7 @@ contract LiquidityStrategy_Test is LiquidityStrategy_BaseTest {
 
     // Should be contraction: add collateral (token1), take debt (token0)
     assertEq(ctx.prices.poolPriceAbove, false);
-    assertEq(uint256(action.dir), uint256(LQ.Direction.Contract));
+    assertEq(action.dir, LQ.Direction.Contract);
     assertTrue(action.amountOwedToPool > 0); // Collateral in
     assertTrue(action.amount0Out > 0); // Debt out
     assertEq(action.amount1Out, 0); // No collateral out
@@ -176,7 +176,7 @@ contract LiquidityStrategy_Test is LiquidityStrategy_BaseTest {
 
     // Should be contraction: take debt (token1) from pool, add collateral (token0)
     assertEq(ctx.prices.poolPriceAbove, true);
-    assertEq(uint256(action.dir), uint256(LQ.Direction.Contract));
+    assertEq(action.dir, LQ.Direction.Contract);
     assertTrue(action.amountOwedToPool > 0); // Collateral in
     assertTrue(action.amount1Out > 0); // Debt out (token1)
     assertEq(action.amount0Out, 0);
@@ -196,7 +196,7 @@ contract LiquidityStrategy_Test is LiquidityStrategy_BaseTest {
 
     // Should be expansion: add debt (token1), take collateral (token0)
     assertEq(ctx.prices.poolPriceAbove, false);
-    assertEq(uint256(action.dir), uint256(LQ.Direction.Expand));
+    assertEq(action.dir, LQ.Direction.Expand);
     assertTrue(action.amountOwedToPool > 0); // Debt in
     assertTrue(action.amount0Out > 0); // Collateral out (token0)
     assertEq(action.amount1Out, 0);
@@ -214,7 +214,7 @@ contract LiquidityStrategy_Test is LiquidityStrategy_BaseTest {
     (, LQ.Action memory action) = strategy.determineAction(address(fpmm));
 
     // Should handle decimal scaling correctly
-    assertEq(uint256(action.dir), uint256(LQ.Direction.Expand));
+    assertEq(action.dir, LQ.Direction.Expand);
     assertTrue(action.amountOwedToPool > 0);
     assertTrue(action.amount1Out > 0);
   }
